Move tool_form_options next to its sole substantive consumer
galaxy/managers/tool_form_options.py contained no manager-shaped code
(no DB session ownership, no app/trans dependence, no model business
logic) — it's response-shaping plumbing for tool-parameter to_dict:
pure utility functions (normalize_pagination, accumulate_with_filter),
dict factories (make_*_entry), and a state holder (DataOptionsBuilder).

Three of four consumers already live under lib/galaxy/tools/; the
fourth is the API route thinly wrapping the tool. Co-locate as
galaxy/tools/parameters/pagination.py.
